+++ .env
@@ -1,5 +1,5 @@
 # FRONTEND LOCAL URL
-# NEXT_PUBLIC_LOCAL_FRONTEND_BASE_URL = http://localhost:3000
+NEXT_PUBLIC_LOCAL_FRONTEND_BASE_URL = http://localhost:3000
 
 # FRONTEND SERVER URL
 NEXT_PUBLIC_FRONTEND_BASE_URL = https://mangoit-lms.mangoitsol.com
@@ -8,7 +8,7 @@ NEXT_PUBLIC_FRONTEND_BASE_URL = https://mangoit-lms.mangoitsol.com
 NEXT_PUBLIC_BASE_URL = https://api-mangoit-lms.mangoitsol.com
 
 # LOCAL API URL
-# NEXT_PUBLIC_LOCAL_BASE_URL = http://localhost:6030
+NEXT_PUBLIC_LOCAL_BASE_URL = http://localhost:6030
 
 NEXT_PUBLIC_AUTH_TOKEN = generatetoken
 NEXT_PUBLIC_REGISTER = registration
