+++ src/config/config.ts
@@ -1,16 +1,15 @@
-//SERVER API URL
-export const BASE_URL = `${process.env.NEXT_PUBLIC_BASE_URL}`;
+// //SERVER API URL
+// export const BASE_URL = `${process.env.NEXT_PUBLIC_BASE_URL}`;
 
-// FRONTEND SERVER URL
-export const FRONTEND_BASE_URL = `${process.env.NEXT_PUBLIC_FRONTEND_BASE_URL}`;
+// // FRONTEND SERVER URL
+// export const FRONTEND_BASE_URL = `${process.env.NEXT_PUBLIC_FRONTEND_BASE_URL}`;
 
-// // FRONTEND LOCAL URL
-// export const FRONTEND_BASE_URL = `${process.env.NEXT_PUBLIC_LOCAL_FRONTEND_BASE_URL}`;
+// FRONTEND LOCAL URL
+export const FRONTEND_BASE_URL = `${process.env.NEXT_PUBLIC_LOCAL_FRONTEND_BASE_URL}`;
 
-// // LOCAL API URL
-// export const BASE_URL = `${process.env.NEXT_PUBLIC_LOCAL_BASE_URL}`;
+// LOCAL API URL
+export const BASE_URL = `${process.env.NEXT_PUBLIC_LOCAL_BASE_URL}`;
 
-console.log(BASE_URL,"BASE_URL")
 export const API = {
   authToken: `${BASE_URL}/${process.env.NEXT_PUBLIC_AUTH_TOKEN}`,
   register: `${BASE_URL}/${process.env.NEXT_PUBLIC_REGISTER}`,
@@ -40,7 +39,7 @@ export const API = {
 
   getSubscription: `${BASE_URL}/${process.env.NEXT_PUBLIC_GET_SUBSCRIPTION}`,
   getSubsById: `${BASE_URL}/${process.env.NEXT_PUBLIC_GET_SUBSCRIPTION_BY_ID}`,
-  
+
   //accept payment
   acceptpayment: `${BASE_URL}/${process.env.NEXT_PUBLIC_ACCEPT_PAYMENT}`,
 };
