+++ src/pages/admin/configurations/emailconfiguration/index.tsx
