+++ src/pages/admin/configurations/stripeconfiguration/index.tsx
