+++ src/pages/admin/courses/allmodules/addmodule.tsx
@@ -7,37 +7,32 @@ import { Box, Button, Card, CardContent, FormControl, Grid, IconButton, InputLab
 import SideBar from "@/common/LayoutNavigations/sideBar";
 import BreadcrumbsHeading from "@/common/BreadCrumbs/breadcrumbs";
 import Footer from "@/common/LayoutNavigations/footer";
-import Navbar from "../../../common/LayoutNavigations/navbar";
+import Navbar from "../../../../common/LayoutNavigations/navbar";
 import RichEditor from "@/common/RichTextEditor/textEditor";
 // Helper Import
 import { Controller, useForm } from "react-hook-form";
 import { yupResolver } from "@hookform/resolvers/yup";
-import { sessionValidations } from '@/validation_schema/sessionValidation';
+import { moduleValidations } from '@/validation_schema/moduleValidation';
 import { LoadingButton } from "@mui/lab";
 import CircularProgressBar from '@/common/CircularProcess/circularProgressBar';
 import SpinnerProgress from '@/common/CircularProgressComponent/spinnerComponent';
 import { capitalizeFirstLetter } from '@/common/CapitalFirstLetter/capitalizeFirstLetter';
 // Types Import
-import { sessionType } from '@/types/sessionType';
 import { courseType } from '@/types/courseType';
 import { moduleType } from '@/types/moduleType';
 // CSS Import
-import styles from "../../../styles/sidebar.module.css";
-import Sessions from "../../../styles/session.module.css"
+import styles from "../../../../styles/sidebar.module.css";
+import ModuleCss from "../../../../styles/modules.module.css"
 import { ToastContainer } from 'react-toastify';
 // API services
 import { HandleCourseGet } from '@/services/course';
-import { HandleModuleGet } from '@/services/module';
-import { HandleSessionCreate } from '@/services/session';
+import { HandleModuleCreate } from '@/services/module';
+
 
 export default function AddSession() {
-
    const router: any = useRouter();
    const [despcriptionContent, setdespcriptionContent] = useState("");
    const [getCourses, setCourses] = useState<courseType | any>();
-   const [getSession, setSession] = useState<sessionType | any>();
-   const [getModules, setModules] = useState<moduleType | any>();
-   const [file, setFile] = useState<string | any>('')
    const [isLoadingButton, setLoadingButton] = useState<boolean>(false);
    const [isLoading, setLoading] = useState<boolean>(false);
 
@@ -47,55 +42,30 @@ export default function AddSession() {
       reset,
       setValue,
       control,
-      formState: { errors }, setError
-   } = useForm<sessionType | any>({
-      resolver: yupResolver(sessionValidations),
+      formState: { errors },
+   } = useForm<moduleType | any>({
+      resolver: yupResolver(moduleValidations),
    });
 
-   const handleContentChange = (value: any, identifier: string) => {
-      // console.log(value, "identifier", identifier, value === '<p><br></p>', value === `<p><br></p>`)
-      if (value === '<p><br></p>') {
-         setError(identifier, { message: 'Description is a required field' });
-      } else {
-         setError(identifier, { message: '' })
-         setValue(identifier, value);
-      }
+   const handleContentChange = (value: string, identifier: string) => {
       setdespcriptionContent(value);
+      setValue(identifier, value);
+
    };
 
    const onSubmit = async (event: any) => {
-      if (errors.description?.message === '') {
-      const reqData: any = {
-         description: event.description,
-         module_id: event.module_id,
-         course_id: event.course_id,
-         title: event.title,
-         attachment: file
-      }
-
-      const formData = new FormData()
-      for (var key in reqData) {
-         formData.append(key, reqData[key]);
-      }
-    
-         setLoading(true);
-         setLoadingButton(false)
+      setLoading(true);
+      setLoadingButton(false)
       try {
-            const res = await HandleSessionCreate(formData)
-            setSession(res.data)
-            setLoading(false);
-            setTimeout(() => {
-               router.push('/courses/allsessions/')
-            }, 1000)
-         } 
-         
-      catch (e) {
+         const res = await HandleModuleCreate(event)
+         setLoading(false);
+         setTimeout(() => {
+            router.push('/courses/allmodules/')
+         }, 1000)
+      } catch (e) {
          console.log(e)
          setLoadingButton(true)
       }
-   }else {
-      setError('description', { message: 'Description is a required field' });
-   }
    };
 
    const getCourseData = () => {
@@ -103,12 +73,6 @@ export default function AddSession() {
          setCourses(courses.data)
       })
    };
-
-   const getModuleData = () => {
-      HandleModuleGet('', '').then((modules) => {
-         setModules(modules.data)
-      })
-   }
 
    useEffect(() => {
       let localData: any;
@@ -118,7 +82,7 @@ export default function AddSession() {
       if (localData) {
          const userId = JSON.parse(localData);
          getCourseData();
-         getModuleData();
+
       }
    }, []);
 
@@ -130,22 +94,7 @@ export default function AddSession() {
       );
    }
 
-   const handleChange = (e: any) => {
-      const file = e.target.files[0];
-
-      if (e.target.name === "attachment") {
-         const reader = new FileReader();
-         reader.onload = (e: any) => {
-            setFile(file);
-            setValue("file", file);
-         }
-         if (file) {
-            reader.readAsDataURL(file);
-         }
-      }
-   }
-
-   // console.log("oopps", errors)
+   // console.log("oopps", getModules)
    return (
       <>
          <Navbar />
@@ -155,9 +104,9 @@ export default function AddSession() {
                {/* breadcumbs */}
                <BreadcrumbsHeading
                   First="Home"
-                  Middle="Session"
-                  Text="SESSION"
-                  Link="/courses/allsessions"
+                  Middle="Module"
+                  Text="MODULE"
+                  Link="/modules/addmodule"
                />
                {/* main content */}
                <Card>
@@ -177,15 +126,15 @@ export default function AddSession() {
                               </Grid>
 
                               <Grid item xs={12} sm={12} md={12} lg={6} >
-                                 <Typography>ADD SESSION</Typography>
-                                 <Grid item xs={12} sm={12} md={12} lg={12} className={Sessions.sessionNameGride} >
+                                 <Typography>ADD MODULE</Typography>
+                                 <Grid item xs={12} sm={12} md={12} lg={12} className={ModuleCss.sessionNameGride} >
 
                                     <Grid item xs={12} sm={12} md={6} lg={6}>
                                        <InputLabel>
-                                          Session Name
+                                          Module Name
                                        </InputLabel>
                                        <TextField
-                                          placeholder="Session Name"
+                                          placeholder="Module Name"
                                           {...register("title")}
                                        />
                                        {errors && errors.title
@@ -194,7 +143,7 @@ export default function AddSession() {
                                     </Grid>
 
                                     <Grid item xs={12} sm={12} md={6} lg={6}>
-                                       <InputLabel>Course of session</InputLabel>
+                                       <InputLabel>Course of Module</InputLabel>
                                        <Controller
                                           name="course_id"
                                           control={control}
@@ -206,7 +155,7 @@ export default function AddSession() {
                                                       Select Course
                                                    </MenuItem>
                                                    {getCourses?.map((course: any) => {
-                                                      return (<MenuItem key={course.course.id} value={course.course.id}>{capitalizeFirstLetter(course?.course.title)}</MenuItem>)
+                                                      return (<MenuItem key={course.id} value={course.id}>{capitalizeFirstLetter(course?.title)}</MenuItem>)
                                                    })}
                                                 </Select>
                                              </FormControl>
@@ -219,28 +168,31 @@ export default function AddSession() {
                                  </Grid>
 
                                  <Grid item xs={12} sm={12} md={12} lg={12} mb={2} >
-                                    <InputLabel>Module of session</InputLabel>
+                                    <InputLabel>Status</InputLabel>
                                     <Controller
-                                       name="module_id"
+                                       name="status"
                                        control={control}
                                        defaultValue=""
                                        render={({ field }) => (
                                           <FormControl fullWidth>
                                              <Select {...field} displayEmpty>
                                                 <MenuItem disabled value="">
-                                                   Select Module
+                                                   Status
                                                 </MenuItem>
-                                                {getModules?.map((module: any) => {
-                                                   return (<MenuItem key={module.module.id} value={module.module.id}>{capitalizeFirstLetter(module?.module.title)}</MenuItem>)
-                                                })}
+                                                <MenuItem value={'active'} >
+                                                   Active
+                                                </MenuItem>
+                                                <MenuItem value={'inactive'}>
+                                                   In-active
+                                                </MenuItem>
                                              </Select>
                                           </FormControl>
                                        )}
                                     />
-                                    {errors && errors.module_id ? ErrorShowing(errors?.module_id?.message) : ""}
+                                    {errors && errors.status ? ErrorShowing(errors?.status?.message) : ""}
                                  </Grid>
 
-                                 <Grid item xs={12} sm={12} md={12} lg={12} mb={2}>
+                                 <Grid item xs={12} sm={12} md={12} lg={12} mb={2} >
                                     <InputLabel>Description</InputLabel>
                                     <RichEditor
                                        {...register("description")}
@@ -249,29 +201,14 @@ export default function AddSession() {
                                           handleContentChange(e, "description")
                                        }
                                     />
-                                    {errors && errors.description ? ErrorShowing(errors?.description?.message) : ""}
-                                    {/* {despcriptionContent ? '' : errors && errors.description ? ErrorShowing(errors?.description?.message) : ""} */}
+                                    {despcriptionContent ? '' : errors && errors.description ? ErrorShowing(errors?.description?.message) : ""}
                                  </Grid>
 
-                                 <Grid item xs={12} sm={12} md={12} lg={12} mb={2}>
-                                    <InputLabel>Attachment</InputLabel>
-                                    <Box className={Sessions.sessionAttachmentBox}>
-                                       <InputLabel className={Sessions.subbox} >
-                                          <input
-                                             type="file"
-                                             {...register('attachment')}
-                                             onChange={handleChange}
-                                             hidden
-                                          />
-                                          <Typography className={Sessions.sessionAttachments}>  {!file.name ? "Upload" : file.name}</Typography>
-                                       </InputLabel>
-                                    </Box>
-                                    {file ? '' : errors && errors.file ? ErrorShowing(errors?.file?.message) : ""}
-                                 </Grid>
+
                                  <Grid item xs={12} sm={12} md={12} lg={12} textAlign={"right"} >
                                     {!isLoadingButton ? <Button type="submit" size="large" variant="contained">
-                                       ADD NEW SESSION
-                                    </Button> : <LoadingButton loading={isLoadingButton} className={Sessions.updateLoadingButton}
+                                       ADD NEW MODULE
+                                    </Button> : <LoadingButton loading={isLoadingButton} className={ModuleCss.updateLoadingButton}
                                        size="large" variant="contained" disabled >
                                        <CircularProgressBar />
                                     </LoadingButton>}
