+++ src/pages/admin/courses/allsessions/updatesession/[id].tsx
@@ -8,7 +8,7 @@ import { Box, Button, Card, CardContent, FormControl, Grid, IconButton, InputLab
 import SideBar from "@/common/LayoutNavigations/sideBar";
 import BreadcrumbsHeading from "@/common/BreadCrumbs/breadcrumbs";
 import Footer from "@/common/LayoutNavigations/footer";
-import Navbar from "../../../../common/LayoutNavigations/navbar";
+import Navbar from "../../../../../common/LayoutNavigations/navbar";
 import RichEditor from "@/common/RichTextEditor/textEditor";
 import Preview from '@/common/previewAttachment';
 // Helper Import
@@ -51,13 +51,13 @@ export default function UpdateSession() {
       register,
       handleSubmit,
       reset,
-      setValue,getValues,
+      setValue, getValues,
       control,
       formState: { errors }, setError
    } = useForm<sessionType | any>({
       resolver: yupResolver(sessionUpdateValidation),
    });
-console.log('getvalue', getValues())
+   console.log('getvalue', getValues())
    const handleContentChange = (value: string, identifier: string) => {
       if (value === '<p><br></p>') {
          setError(identifier, { message: 'Description is a required field' });
@@ -71,11 +71,11 @@ console.log('getvalue', getValues())
    };
 
    const onSubmit = async (event: any) => {
-     
-      const id = router.query.id 
-   
+
+      const id = router.query.id
+
       // const reqData = { ...event, 'attachment': file }
-      if (errors.description?.message === '' || ( typeof errors === 'object' && errors !== null)) {
+      if (errors.description?.message === '' || (typeof errors === 'object' && errors !== null)) {
          const reqData: any = {
             description: event.description,
             module_id: event.module_id,
