+++ src/pages/admin/dashboard/index.tsx
@@ -1,7 +1,7 @@
 import Navbar from "@/common/LayoutNavigations/navbar";
 import SideBar from "@/common/LayoutNavigations/sideBar";
 import { Box } from "@mui/material";
-import styles from '../../styles/sidebar.module.css'
+import styles from '../../../styles/sidebar.module.css'
 import BreadcrumbsHeading from "@/common/BreadCrumbs/breadcrumbs";
