+++ src/pages/admin/user/subscription/index.tsx
@@ -20,17 +20,17 @@ import {
   Typography,
 } from "@mui/material";
 // CSS Import
-import CloseIcon from '@mui/icons-material/Close';
-import styles from "../../../styles/sidebar.module.css";
-import courseStyle from "../../../styles/course.module.css";
+import CloseIcon from "@mui/icons-material/Close";
+import styles from "../../../../styles/sidebar.module.css";
+import courseStyle from "../../../../styles/course.module.css";
 import BreadcrumbsHeading from "@/common/BreadCrumbs/breadcrumbs";
 import Footer from "@/common/LayoutNavigations/footer";
 import PopupState, { bindTrigger, bindPopover } from "material-ui-popup-state";
 import FilterAltOutlinedIcon from "@mui/icons-material/FilterAltOutlined";
-import DeleteOutlineIcon from '@mui/icons-material/DeleteOutline';
-import ModeEditOutlineIcon from '@mui/icons-material/ModeEditOutline';
-import ArrowDownwardOutlinedIcon from '@mui/icons-material/ArrowDownwardOutlined';
-import ArrowUpwardOutlinedIcon from '@mui/icons-material/ArrowUpwardOutlined';
+import VisibilityIcon from "@mui/icons-material/Visibility";
+import ModeEditOutlineIcon from "@mui/icons-material/ModeEditOutline";
+import ArrowDownwardOutlinedIcon from "@mui/icons-material/ArrowDownwardOutlined";
+import ArrowUpwardOutlinedIcon from "@mui/icons-material/ArrowUpwardOutlined";
 import * as React from "react";
 import Paper from "@mui/material/Paper";
 import Table from "@mui/material/Table";
@@ -42,15 +42,16 @@ import TablePagination from "@mui/material/TablePagination";
 import TableRow from "@mui/material/TableRow";
 import { SearchOutlined } from "@mui/icons-material";
 import { useRouter } from "next/router";
-import { HandleCourseDelete, HandleCourseGet } from "@/services/course";
+import { HandleSubscriptionGet } from "@/services/subscription";
 import { capitalizeFirstLetter } from "@/common/CapitalFirstLetter/capitalizeFirstLetter";
 import { usePagination } from "@/common/Pagination/paginations";
 import { AlertDialog } from "@/common/DeleteListRow/deleteRow";
-import { Controller,useForm } from "react-hook-form";
+import { Controller, useForm } from "react-hook-form";
 import { handleSortData } from "@/common/Sorting/sorting";
+import Link from "next/link";
 
 interface Column {
-  id: "id" | "title" | "module" | "session" | "is_chargeable" | "status" | "action";
+  id: "id" | "amount" | "next_pay" | "status" | "action";
   label: string;
   minWidth?: number;
   align?: "right";
@@ -58,26 +59,23 @@ interface Column {
 }
 
 const columns: Column[] = [
-  { id: "id", label: "ID", },
-  { id: "title", label: "COURSE NAME", minWidth: 170 },
-  { id: "module", label: "NO. MODULE", minWidth: 100 },
-  { id: "session", label: "NO. SESSION", minWidth: 100 },
-  { id: "is_chargeable", label: "TYPE", minWidth: 100 },
+  { id: "id", label: "ID", minWidth: 100 },
+  { id: "amount", label: "AMOUNT", minWidth: 100 },
+  { id: "next_pay", label: "NEXT PAY", minWidth: 100 },
   { id: "status", label: "STATUS", minWidth: 100 },
   { id: "action", label: "ACTION", minWidth: 100 },
 ];
 
-const AllCourses = () => {
+const Subscription = () => {
   const [rows, setRows] = React.useState<any>([]);
   const [toggle, setToggle] = React.useState<boolean>(false);
-  const [search, setSearch] = React.useState('');
-  const [deleteRow, setDeleteRow] = React.useState<any>([])
+  const [search, setSearch] = React.useState("");
+  const [deleteRow, setDeleteRow] = React.useState<any>([]);
   const [open, setOpen] = React.useState(false);
   const [getFilter, setFilter] = React.useState<number>(0);
-  const [filterObject, setFilterObject] = React.useState<any>('');
-  
+  const [filterObject, setFilterObject] = React.useState<any>("");
 
-  const router = useRouter()
+  const router = useRouter();
   //pagination
   const [row_per_page, set_row_per_page] = React.useState(5);
   let [page, setPage] = React.useState<any>(1);
@@ -92,74 +90,42 @@ const AllCourses = () => {
     DATA.jump(p);
   };
 
-  const {
-    handleSubmit,
-    control,
-    reset,
-  } = useForm();
+  const handleClickOpen = (id: any) => {
+    router.push(`/user/subscription/view/${id}`);
+  };
 
-  const onSubmit = (event: any) => {
-    HandleCourseGet('', event).then((itemFiltered) => {
-      setRows(itemFiltered.data)
-      setFilterObject(event)
-    })
-  }
-  
-
-  const handleClickOpen = (row: any) => {
-    // console.log('row', row)
-    setDeleteRow(row)
-    setOpen(!open);
-
-  }
-  // to delete a row
-  const handleDeletesRow = () => {
-    HandleCourseDelete(deleteRow.id).then((deletedRow) => {
-      HandleCourseGet('', filterObject).then((newRows) => {
-        setRows(newRows.data)
-      })
-    })
-    setOpen(!open);
-  }
-
-  const resetFilterValue = () => {
-    setFilter(0)
-    reset({ type: 0, status: 0 });
-  }
   const handleSort = (rowsData: any) => {
-    const sortData = handleSortData(rowsData)
-    setRows(sortData)
-    setToggle(!toggle)
-  }
-  const handleSearch = (e: any, identifier: any) => {
-    setPage(1);
-    DATA.jump(1);
-    if (identifier === 'reset') {
-      HandleCourseGet('', { type: 0, status: 0 }).then((itemSeached) => {
-        setRows(itemSeached.data);
-      })
-      setSearch(e)
-    } else {
-      const search = e.target.value;
-      setSearch(e.target.value)
-      HandleCourseGet(search, filterObject).then((itemSeached) => {
-        setRows(itemSeached.data);
-      })
-    }
-  }
+    const sortData = handleSortData(rowsData);
+    setRows(sortData);
+    setToggle(!toggle);
+  };
+  //   const handleSearch = (e: any, identifier: any) => {
+  //     setPage(1);
+  //     DATA.jump(1);
+  //     if (identifier === "reset") {
+  //       HandleCourseGet("", { type: 0, status: 0 }).then((itemSeached) => {
+  //         setRows(itemSeached.data);
+  //       });
+  //       setSearch(e);
+  //     } else {
+  //       const search = e.target.value;
+  //       setSearch(e.target.value);
+  //       HandleCourseGet(search, filterObject).then((itemSeached) => {
+  //         setRows(itemSeached.data);
+  //       });
+  //     }
+  //   };
 
   const getAllCourseData = () => {
-    HandleCourseGet('', filterObject).then((courses) => {
-      console.log(courses,"45")
-      setRows(courses.data)
-    })
-  }
+    HandleSubscriptionGet().then((subs) => {
+      setRows(subs.data);
+    });
+  };
 
   React.useEffect(() => {
     getAllCourseData();
-  }, [])
+  }, []);
 
-  // console.log('rowww', rows)
   return (
     <>
       <Navbar />
@@ -170,15 +136,15 @@ const AllCourses = () => {
           {/* breadcumbs */}
           <BreadcrumbsHeading
             First="Home"
-            Middle="Courses"
-            Text="COURSES"
-            Link="/courses/allcourses"
+            Middle="Subscription"
+            Text="SUBSCRIPTION"
+            Link="/user/subscription"
           />
 
           {/* main content */}
           <Card>
             <CardContent>
-              <TextField
+              {/* <TextField
                 id="standard-search"
                 value={search}
                 variant="outlined"
@@ -191,8 +157,8 @@ const AllCourses = () => {
                     </IconButton> : <IconButton onClick={(e) => handleSearch('', 'reset')}> <CloseIcon /></IconButton>
                   ),
                 }}
-              />
-              <Box
+              /> */}
+              {/* <Box
                 sx={{ float: "right", display: "flex", alignItems: "center" }}
               >
                 <PopupState variant="popover" popupId="demo-popup-popover" >
@@ -292,7 +258,7 @@ const AllCourses = () => {
                                       lg={12}
                                     >
                                   <Box >
-                                        <Button className={courseStyle.boxInFilter}
+                                        <Button
                                           size="medium"
                                           variant="contained"
                                           color="primary"
@@ -306,7 +272,7 @@ const AllCourses = () => {
                                           type="submit"
                                           variant="contained"
                                           color="primary"
-                                          className={courseStyle.applyButtonInFiltter}
+                                        
                                           onClick={popupState.close}
                                         >
                                           Apply
@@ -325,8 +291,8 @@ const AllCourses = () => {
                 </PopupState>
                 &nbsp;
                 <Button variant="contained" onClick={() => router.push('/courses/allcourses/addcourse')}>Add New Course</Button>
-              </Box>
-              <Paper >
+              </Box> */}
+              <Paper>
                 <TableContainer className={courseStyle.tableContainer}>
                   <Table stickyHeader aria-label="sticky table">
                     <TableHead>
@@ -337,41 +303,84 @@ const AllCourses = () => {
                             align={column.align}
                             style={{ top: 0, minWidth: column.minWidth }}
                             onClick={() => {
-                              column.label === 'ID' ?
-                                handleSort(rows) :
-                                ''
+                              column.label === "ID" ? handleSort(rows) : "";
                             }}
                           >
-                            {toggle ? column.label === 'ID' ? <Typography>ID <ArrowDownwardOutlinedIcon fontSize="small" /> </Typography> : column.label : column.label === 'ID' ? <Typography>ID <ArrowUpwardOutlinedIcon fontSize="small" /> </Typography> : column.label}
+                            {toggle ? (
+                              column.label === "ID" ? (
+                                <Typography>
+                                  ID{" "}
+                                  <ArrowDownwardOutlinedIcon fontSize="small" />{" "}
+                                </Typography>
+                              ) : (
+                                column.label
+                              )
+                            ) : column.label === "ID" ? (
+                              <Typography>
+                                ID <ArrowUpwardOutlinedIcon fontSize="small" />{" "}
+                              </Typography>
+                            ) : (
+                              column.label
+                            )}
                           </TableCell>
                         ))}
                       </TableRow>
                     </TableHead>
                     <TableBody>
-                      {console.log(rows,rows?.length,DATA)}
-                      {rows && rows.length > 0 ? DATA.currentData() &&
-                        DATA.currentData()
-                          .map((row: any) => {
-                            const statusColor = (row.course.status === "active" ? courseStyle.activeClassColor : row.course.status === "inactive" ? courseStyle.inactiveClassColor : courseStyle.draftClassColor)
-                            return (
-                              <TableRow
-                                hover
-                                role="checkbox"
-                                tabIndex={-1}
-                                key={row.id}
-                              >
-                                <TableCell>{row.course.id}</TableCell>
-                                <TableCell>{capitalizeFirstLetter(row?.course?.title)}</TableCell>
-                                <TableCell>{row?.moduleCount?.length !== 0 ? row?.moduleCount[0]?.moduleCount : 0}</TableCell>
-                                <TableCell>{row?.sessionCount?.length !== 0 ? row?.sessionCount[0]?.sessionCount : 0}</TableCell>
-                                <TableCell>{capitalizeFirstLetter(row?.course?.is_chargeable)}</TableCell>
-                                <TableCell className={statusColor}>{capitalizeFirstLetter(row?.course?.status)}</TableCell>
-                                <TableCell><Button onClick={() => router.push(`/courses/allcourses/updatecourse/${row.course.id}`)} variant="outlined" color="success" className={courseStyle.editDeleteButton}  ><ModeEditOutlineIcon /></Button>
-                                  <Button className={courseStyle.editDeleteButton} variant="outlined" color="error" onClick={() => handleClickOpen(row?.course)}><DeleteOutlineIcon /></Button>
-                                </TableCell>
-                              </TableRow>
-                            );
-                          }) : <TableRow><TableCell colSpan={7} sx={{fontWeight:600,textAlign:'center'}}> Record not Found </TableCell></TableRow>}
+                      {rows && rows.length > 0 ? (
+                        DATA.currentData() &&
+                        DATA.currentData().map((row: any) => {
+                          let color = "active";
+                          const statusColor =
+                            color === "active"
+                              ? courseStyle.activeClassColor
+                              : color === "inactive"
+                              ? courseStyle.inactiveClassColor
+                              : courseStyle.draftClassColor;
+                          return (
+                            <TableRow
+                              hover
+                              role="checkbox"
+                              tabIndex={-1}
+                              key={row.id}
+                            >
+                              <TableCell>{row?.id}</TableCell>
+                              <TableCell>${row?.price}</TableCell>
+                              <TableCell>
+                                {/* {capitalizeFirstLetter(
+                                  row?.course?.is_chargeable
+                                )} */}
+                                25 May 2023
+                              </TableCell>
+                              <TableCell className={statusColor}>
+                                {/* {capitalizeFirstLetter(row?.course?.status)} */}
+                                Active
+                              </TableCell>
+                              <TableCell>
+                                <Button
+                                  className={courseStyle.editDeleteButton}
+                                  //   href="/user/subscription/view"
+                                  variant="outlined"
+                                  color="primary"
+                                  onClick={() => handleClickOpen(row?.id)}
+                                >
+                                  <VisibilityIcon />
+                                </Button>
+                              </TableCell>
+                            </TableRow>
+                          );
+                        })
+                      ) : (
+                        <TableRow>
+                          <TableCell
+                            colSpan={7}
+                            sx={{ fontWeight: 600, textAlign: "center" }}
+                          >
+                            {" "}
+                            Record not Found{" "}
+                          </TableCell>
+                        </TableRow>
+                      )}
                     </TableBody>
                   </Table>
                   <Stack
@@ -394,7 +403,7 @@ const AllCourses = () => {
                         defaultValue={5}
                         onChange={handlerowchange}
                         size="small"
-                        style={{ height: "40px", marginRight: '11px' }}
+                        style={{ height: "40px", marginRight: "11px" }}
                       >
                         <MenuItem value={5}>5</MenuItem>
                         <MenuItem value={20}>20</MenuItem>
@@ -404,13 +413,13 @@ const AllCourses = () => {
                   </Stack>
                 </TableContainer>
               </Paper>
-              <AlertDialog
+              {/* <AlertDialog
                 open={open}
                 onClose={handleClickOpen}
                 onSubmit={handleDeletesRow}
                 title={deleteRow.title}
-                whatYouDelete='Session'
-              />
+                whatYouDelete="Session"
+              /> */}
             </CardContent>
           </Card>
         </Box>
@@ -420,4 +429,4 @@ const AllCourses = () => {
   );
 };
 
-export default AllCourses;
+export default Subscription;
