+++ src/styles/payment.module.css
@@ -45,7 +45,7 @@
 
 @media (min-width: 360px) {
   .h1 {
-    font-size: 2.5em;
+    font-size: 2rem;
   }
   .gohome {
     margin-bottom: 0px;
@@ -95,8 +95,5 @@
     display: flex;
     align-items: center;
     justify-content: center;
-  }
-  .h1 {
-    font-size: "10px";
   }
 }
