+++ src/styles/webview.module.css
@@ -1,0 +1,329 @@
+/*landing page*/
+/* ==== WRAPPER ==== */
+.wrapper {
+  width: 100%;
+  padding-top: 2rem;
+  margin-bottom: 2rem;
+}
+.container {
+  width: 100%;
+}
+
+.gridcols2 {
+  width: 100%;
+  height: 100%;
+  display: grid;
+  grid-template-columns: repeat(2, 1fr);
+  gap: 4rem;
+}
+
+.griditem1 {
+  padding-top: 5rem;
+  padding-left: 1.6rem;
+}
+
+.mainheading {
+  font-weight: 800;
+  font-size: 40px;
+  line-height: 55px;
+}
+
+.mainheading span {
+  font-weight: 800;
+  font-size: 40px;
+  line-height: 55px;
+  color: rgb(236, 113, 41);
+}
+.infotext {
+  margin-top: 1.5rem;
+  font-size: 16px;
+  line-height: 25px;
+  color: #334157;
+}
+
+.btnwrapper {
+  margin-top: 1.5rem;
+  display: flex;
+  width: 100%;
+}
+
+.btnwrapper {
+  width: 110px;
+  height: 50px;
+  background-color: #646262;
+  display: block;
+  font-size: 16px;
+  color: #fff;
+  text-transform: capitalize;
+  border-radius: 7px;
+  letter-spacing: 1px;
+  transition: 0.4s;
+}
+.btnwrapper:hover {
+  transform: translateY(-3px);
+}
+
+.viewmorebtn {
+  width: 180px;
+  height: 55px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  font-size: 16px;
+  letter-spacing: 0;
+  color: #fff;
+  font-weight: 500;
+  margin-right: 10px;
+}
+
+.viewmorebtn:hover {
+  transition: box-shadow 0.25s ease, transform 0.25s ease;
+}
+
+.griditem2 {
+  width: 100%;
+  height: 100%;
+}
+
+.teamimgwrapper {
+  width: 520px;
+  max-width: 100%;
+  height: 440px;
+}
+
+.teamimgwrapper img {
+  width: 100%;
+  height: 100%;
+  object-fit: contain;
+}
+
+@media screen and (max-width: 768px) {
+  .wrapper {
+    padding: 0 0.7rem;
+  }
+  .griditem1 {
+    padding-left: 0rem;
+  }
+  .mainheading {
+    font-size: 35px;
+  }
+  .viewmorebtn {
+    width: 140px;
+    height: 55px;
+    font-size: 13.5px;
+    margin-right: 1rem;
+  }
+}
+
+@media screen and (max-width: 991px) {
+  .wrapper {
+    padding-top: 3rem;
+  }
+  .gridcols2 {
+    grid-template-columns: repeat(auto-fit, minmax(100%, 1fr));
+  }
+  .griditem1 {
+    order: 2;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    padding-top: 0;
+  }
+  .mainheading {
+    font-size: 32px;
+    text-align: center;
+    line-height: 40px;
+  }
+  .infotext {
+    font-size: 16px;
+    text-align: center;
+    padding: 0.7rem;
+  }
+  .btnwrapper {
+    width: 100%;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+  }
+  .griditem2 {
+    order: 1;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+  }
+  .teamimgwrapper {
+    width: 350px;
+    height: 350px;
+  }
+}
+
+/**/
+/*icons cards*/
+.iconcard {
+  padding-top: 2rem;
+  margin-bottom: 2rem;
+}
+
+.iconcardarticles {
+  --img-scale: 1.001;
+  --title-color: black;
+  --link-icon-translate: -20px;
+  --link-icon-opacity: 0;
+  position: relative;
+  border-radius: 16px;
+  box-shadow: none;
+  background: #fff;
+  transform-origin: center;
+  transition: all 0.4s ease-in-out;
+  overflow: hidden;
+}
+
+.articles {
+  --img-scale: 1.001;
+  --title-color: black;
+  --link-icon-translate: -20px;
+  --link-icon-opacity: 0;
+  position: relative;
+  border-radius: 16px;
+  box-shadow: none;
+  background: #fff;
+  transform-origin: center;
+  transition: all 0.4s ease-in-out;
+  overflow: hidden;
+}
+
+.articles {
+  display: grid;
+  max-width: 1200px;
+  margin-inline: auto;
+  padding-inline: 24px;
+  grid-template-columns: repeat(auto-fill, minmax(250px, 1fr));
+  gap: 24px;
+}
+
+.article {
+  --img-scale: 1.001;
+  --title-color: black;
+  --link-icon-translate: -20px;
+  --link-icon-opacity: 0;
+  position: relative;
+  border-radius: 16px;
+  box-shadow: none;
+  background: #fff;
+  transform-origin: center;
+  transition: all 0.4s ease-in-out;
+  overflow: hidden;
+}
+
+.article a::after {
+  position: absolute;
+  inset-block: 0;
+  inset-inline: 0;
+  cursor: pointer;
+  content: "";
+}
+
+/* basic article elements styling */
+.article .h2 {
+  margin: 0 0 18px 0;
+  font-family: "Bebas Neue", cursive;
+  font-size: 1.9rem;
+  letter-spacing: 0.06em;
+  color: var(--title-color);
+  transition: color 0.3s ease-out;
+}
+
+.figure {
+  margin: 0;
+  padding: 0;
+  aspect-ratio: 16 / 9;
+  overflow: hidden;
+}
+
+.article img {
+  max-width: 100%;
+  transform-origin: center;
+  transform: scale(var(--img-scale));
+  transition: transform 0.4s ease-in-out;
+}
+
+.articlebody {
+  padding: 24px;
+}
+
+.article a {
+  display: inline-flex;
+  align-items: center;
+  text-decoration: none;
+  color: #28666e;
+}
+
+.article a:focus {
+  outline: 1px dotted #28666e;
+}
+
+.article a .icon {
+  min-width: 24px;
+  width: 24px;
+  height: 24px;
+  margin-left: 5px;
+  transform: translateX(var(--link-icon-translate));
+  opacity: var(--link-icon-opacity);
+  transition: all 0.3s;
+}
+
+/* using the has() relational pseudo selector to update our custom properties */
+.article:has(:hover, :focus) {
+  --img-scale: 1.1;
+  --title-color: #28666e;
+  --link-icon-translate: 0;
+  --link-icon-opacity: 1;
+  box-shadow: rgba(0, 0, 0, 0.16) 0px 10px 36px 0px,
+    rgba(0, 0, 0, 0.06) 0px 0px 0px 1px;
+}
+
+/************************ 
+  Generic layout (demo looks)
+  **************************/
+
+@media screen and (max-width: 960px) {
+  .article {
+    container: card/inline-size;
+  }
+  .articlebody p {
+    display: none;
+  }
+}
+
+@container card (min-width: 380px) {
+  .articlewrapper {
+    display: grid;
+    grid-template-columns: 100px 1fr;
+    gap: 16px;
+  }
+  .articlebody {
+    padding-left: 0;
+  }
+  .figure {
+    width: 100%;
+    height: 100%;
+    overflow: hidden;
+  }
+  .figure img {
+    height: 100%;
+    aspect-ratio: 1;
+    object-fit: cover;
+  }
+}
+
+.sr-only:not(:focus):not(:active) {
+  clip: rect(0 0 0 0);
+  clip-path: inset(50%);
+  height: 1px;
+  overflow: hidden;
+  position: absolute;
+  white-space: nowrap;
+  width: 1px;
+}
